feat!: remove ECX resources (#683)
The ECX v3 API will be shut down at the end of June 2024. This removes
the corresponding Terraform resources and datasources, as well as any
supporting code that was only used by those resources and datasources.

This is a breaking change that will trigger a major version bump. Any
users who have not transitioned from ECX resources to Fabric resources
will still be able to use `1.x` releases of the terraform provider to
manage ECX resources until the ECX v3 API is turned off.