feat: Support IPv6 parameters when creating / updating interconnection virtual circuits (#717)
This PR adds support for the newly available IPv6 parameters when
creating / updating interconnection virtual circuits. To allow this
feature to also be used with shared interconnections, some changes to
the create logic of the interconnection virtual circuit resource were
made.

With shared interconnections, the customer does not create virtual
circuits, but rather they are created for them by the metal API.
Attempting to create a virtual circuit on a shared connection will
result in an error from the Metal API.

To get around this, an optional `virtual_circuit_id` parameter has been
added to the virtual circuit resource. If this value is present, the
terraform create function will issue an update call to the metal API for
the existing specified virtual circuit, rather than trying to create a
new one.

Additionally I have added support for the already existing IPv4
parameters in the update function which seemed to be missing.