chore(ci): lint with golangci-lint (#726)
This introduces linting in CI to help maintain code quality and identify
issues that may not be caught by tests. To avoid a big PR that resolves
all existing linting issues, the CI workflow is configured to only look
at files that are changed in a particular PR or commit; we'll need to
keep an eye on this to make sure it's working as expected. This will
allow us to chip away at existing issues over time. Eventually (ideally
within a few months), the CI workflow should be updated to always lint
the entire codebase.

This PR also introduces a `make lint` task for local execution, but this
task lints the entire codebase and, for the time being, will catch some
issues that are not flagged for a particular PR or commit in CI.