feat: remove credential validation on provider startup
Previously we validated that the user specified at least one kind
of Equinix credential during provider startup.  This behavior is
unusual among Terraform providers and its utility is debatable,
because it does not and cannot protect against these situations:

- Invalid credentials were provided
- Credentials were omitted for a service that is used (e.g., a Metal
  auth token was specified, but a client ID and secret were needed
  for the resources in the user's configuration)

In addition to the above shortcomings, validating credentials in
the Config type that is shared between SDKv2 and framework providers
forces unnecessary changes to the SDKv2 provider if a new service is
onboarded to the framework provider.

This removes the credential validation from the config object so
that any and all validation of the provider configuration is specified
in the provider schema.