chore: prevent new Go files from being added to equinix package (#808)
To ensure that we continue to make progress on #106, this updates the
`Validate PR` GitHub Actions Workflow to check if new Go files have been
added to the `equinix/` directory.

The workflow only checks for [files with `"added"`
status](https://docs.github.com/en/rest/pulls/pulls?apiVersion=2022-11-28#list-pull-requests-files)
that are directly under the `equinix/` directory. This leaves room for
renaming/copying files and for adding new packages in
`equinix/<subdirectory>/`.